Moved all UI code into one feature module
