+++ base/src/main/AndroidManifest.xml
@@ -13,26 +13,6 @@
         <meta-data
             android:name="aia-compat-api-min-version"
             android:value="1" />
-
-        <activity android:name="com.prassatyan.resume.MainActivity">
-            <intent-filter android:order="1">
-                <action android:name="android.intent.action.VIEW" />
-
-                <category android:name="android.intent.category.BROWSABLE" />
-                <category android:name="android.intent.category.DEFAULT" />
-
-                <data
-                    android:host="praszapps.com"
-                    android:pathPattern="/.*"
-                    android:scheme="https" />
-            </intent-filter>
-
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
         <activity android:name=".ui.ResumeBaseActivity" />
     </application>
