+++ resumefeature/src/main/AndroidManifest.xml
@@ -1,3 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.prassatyan.resume.feature.resumemain"/>
+    package="com.prassatyan.resume.feature.resumemain">
+
+    <application>
+        <activity android:name="com.prassatyan.resume.feature.resumemain.MainActivity">
+            <intent-filter android:order="1">
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.BROWSABLE" />
+                <category android:name="android.intent.category.DEFAULT" />
+
+                <data
+                    android:host="praszapps.com"
+                    android:pathPattern="/.*"
+                    android:scheme="https" />
+            </intent-filter>
+
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>
