+++ resumefeature/src/main/java/com/prassatyan/resume/feature/resumemain/GithubFragment.java
@@ -1,10 +1,11 @@
-package com.prassatyan.resume.feature.github;
+package com.prassatyan.resume.feature.resumemain;
 
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
 
@@ -23,6 +24,8 @@ public class GithubFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
+
+        Toast.makeText(getContext(), "Clicked on github", Toast.LENGTH_SHORT).show();
         return inflater.inflate(R.layout.fragment_github, container, false);
     }
